requirements.txt
backtesting~=0.3.3
bitmex~=0.2.2
prophet~=1.1.5
hydra-core~=1.3.2
jsonlines~=4.0.0
jupyter~=1.0.0
keras~=2.15.0
matplotlib~=3.8.4
neuralprophet~=0.7.0
numpy~=1.26.4
omegaconf~=2.3.0
orbit-ml~=1.1.4
pandas~=2.2.2
pillow~=10.3.0
scikit-learn~=1.4.2
scipy~=1.11.4
seaborn~=0.13.2
statsmodels~=0.14.2
tensorflow~=2.15.0
tqdm~=4.66.4
xgboost~=2.0.3
plotly~=5.22.0
streamlit~=1.36.0
python-telegram-bot~=20.8
